\name{exmpDat}
\alias{exmpDat}
\docType{data}
\title{
Example Data
}
\description{
exmpDat
}
\usage{data("exmpDat")}
\format{
  A data frame with 191 observations on the following 5 variables.
  \describe{
    \item{\code{ID}}{subject ID}
    \item{\code{street}}{a character vector of street address}
    \item{\code{city}}{a character vector of city in US}
    \item{\code{state}}{a character vector of state in US}
    \item{\code{zipcode}}{a vector of zip code in formt of 5 digits}
    \item{\code{Address}}{a character vector of address in format of street, city, state, zipcode}
  }
}
\details{
exmpDat
}
\examples{
data(exmpDat)
head(exmpDat)
}
\keyword{datasets}
