\name{ezADI-package}
\alias{ezADI-package}
\docType{package}
\title{
\packageTitle{ezADI}
}
\description{
\packageDescription{ezADI}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ezADI}
\packageIndices{ezADI}
An overview of how to use the package, including the most important functions
}
\author{
\packageAuthor{ezADI}

Maintainer: \packageMaintainer{ezADI}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{\code{\link{ezADI}}}
\examples{
data(exmpDat)
test<- ezADI(inputFile = NULL, inputDat = exmpDat, adiFile = "US_ADI_Example.csv", outputFile = NULL, batchSize = 100)
}
